fix Whisper FP16/FP32 dtype mismatch, fix wav2vec CTC decode
- Whisper: load with torch_dtype=float32 (checkpoint is FP16 by default)
- IndicWav2Vec: use tokenizer.decode() instead of processor.batch_decode()
  (processor expects logits for LM decoding, not token IDs)

+++ src/bhai/stt/vaani_whisper.py
@@ -24,10 +24,13 @@ class VaaniWhisperSTT(GPUModelSTT):
         return self.model_id
 
     def _load_model(self) -> None:
+        import torch
         from transformers import WhisperForConditionalGeneration, WhisperProcessor
 
         self._processor = WhisperProcessor.from_pretrained(self.model_id)
-        self._model = WhisperForConditionalGeneration.from_pretrained(self.model_id)
+        self._model = WhisperForConditionalGeneration.from_pretrained(
+            self.model_id, torch_dtype=torch.float32
+        )
         self._model.to(self.device)
 
     def transcribe(self, audio_path: Path) -> Dict[str, Any]:
